requirements.txt
fastapi
rembg
uvicorn
pillow
